update and delete a service !!

+++ htdocs/zabbix/web/app/controllers/action.php
@@ -2,7 +2,6 @@
 
 	require('../database/connection.php');	
 	require('../models/value_sla.php');
-	require('../models/service.php');
 	require('../models/host_service.php');
 	require('../models/sla_status.php');
 	require('serviceController.php');
@@ -12,15 +11,48 @@
 		switch($action)
 		{
 			case 'delete_service' :	$serviceController = new serviceController();
-									$test = $serviceController->delete($_POST['id']);
-									if($test)
+									if(isset($_POST['id']))
 									{
-										echo "good";
+										$test = $serviceController->delete($_POST['id']);
+										if($test)
+										{
+											echo "good";
+										}
+										else
+										{
+											echo "bad";
+										}	
 									}
 									else
 									{
 										echo "bad";
 									}
+									
+			break;
+			case 'update_service' : $serviceController = new serviceController();
+									$host =new Host_Service();
+									
+									if((isset($_POST['id'])) && (isset($_POST['host_name'])) && (isset($_POST['ip_address'])))
+									{
+										$serviceController = new serviceController();
+										$h = new Host_Service();
+										$h->setHost_name($_POST['host_name']);
+										$h->setIp_address($_POST['ip_address']);
+										$test = $serviceController->update($_POST['id'],$h);
+										if($test)
+										{
+											echo "good";
+										}
+										else
+										{
+											echo "bad";
+										}
+									}
+									else
+									{
+										echo "bad";
+									}
+									
 			break;
 		}
 	}
